Shift the createSingleUser and createMultipleUsers out
