--- src/server/trpc/router/users/auth.ts
@@ -1,6 +1,0 @@
-import { router } from '../../trpc'
-import { resetPassword } from './post'
-
-export const authRouter = router({
-  resetPassword,
-})
